Start explicit transaction
By default, MySQLdb disables autocommit, but when
reconnecting with connection.ping(True), autocommit
is not disabled.

The two options are: (a) set autocommit to off, or
(b) make an explicit begin for the multiple statement
transaction by using connection.begin().

I prefer (b) because it is clear what the code is doing
and does not rely on current session config.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>